refactor(api): create service for user certificability check
